address various review comments
This commit will address various review comments on #77:

- Move TestEntityTagMarshalJSON into types/entity_test.go
- Delete entity_tag_test.go
- Remove "omitempty" JSON tag from Entity.Tag field
- Rename "UnspecifiedEntity" -> "ZeroEntity" in eval/tests
- Rename "nil-entity-map" -> "nil-entity-getter" and explicitly
  set Entities: nil in TestIsAuthorized
- Move a tag authorization test from corpus_test.go
  to authorize_test.go
- Uncomment a test directive which was commented by mistake
